infrastructure/requirements.txt
# AWS SDK
boto3>=1.34.0

# AWS CDK
aws-cdk-lib==2.215.0
constructs>=10.0.0,<11.0.0
